Refine comments and seal TestDataChild class
In `ExtensionsTestsDataSources.cs`, comments were refined to improve clarity and correctness. The comments now explicitly state that the object array returned is the same object array with the new null element.

In `TestDataChild.cs`, the `TestDataChild` class was modified from a `record` to a `sealed record`, which means that this class is now sealed and cannot be inherited by other classes.