Make ParamName and Message properties nullable
The changes made to the `ITestDataThrows.cs` and `TestDataThrows.cs` files involve making the `ParamName` and `Message` properties nullable. This is indicated by the addition of the `?` symbol to the type declarations of these properties. Specifically, the `ParamName` property in the `ITestDataThrows` interface and the `TestDataThrows` record was changed from `string` to `string?`, and the `Message` property was similarly changed from `string` to `string?`.