Add unit test for ToString method in TestDataTests
Added a new test method `ToString_returnsExpected` to the `TestDataTests` class in the `TestDataTests.cs` file. This unit test verifies that the `ToString` method of the `_sut` object returns the expected string representation, which is the value of `_sut.TestCase`. This ensures the correctness of the `ToString` implementation for the `TestData` class.